+++ distribution-lifecycle/src/test/java/io/telicent/smart/cache/distribution/lifecycle/events/listeners/TestAcknowledgingListener.java
@@ -24,6 +24,7 @@ import io.telicent.smart.cache.distribution.lifecycle.store.apps.AppDistribution
 import io.telicent.smart.cache.distribution.lifecycle.tracker.TemporarilyFails;
 import io.telicent.smart.cache.payloads.Envelope;
 import io.telicent.smart.cache.payloads.LazyEnvelope;
+import io.telicent.smart.cache.projectors.Sink;
 import io.telicent.smart.cache.projectors.sinks.CollectorSink;
 import io.telicent.smart.cache.sources.Event;
 import org.testng.Assert;
@@ -38,14 +39,17 @@ import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import static io.telicent.smart.cache.distribution.lifecycle.Util.ack;
 import static io.telicent.smart.cache.distribution.lifecycle.Util.action;
-import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.*;
 
 // java:S2925 - Thread.sleep is required when waiting on real Kafka/Docker in integration tests
 @SuppressWarnings("java:S2925")
 public class TestAcknowledgingListener {
 
     public static final String APP_ID = "test";
+
+    public static final String DISTRO_ID = "https://example.org/distribution/test";
 
     private static final class Ok implements DistributionLifecycleListener {
 
@@ -102,7 +106,8 @@ public class TestAcknowledgingListener {
             AcknowledgingListener listener = AcknowledgingListener.builder()
                                                                   .listener(new Ok())
                                                                   .sink(collector)
-                                                                  .stateStore(mock(DistributionLifecycleStateStore.class))
+                                                                  .stateStore(
+                                                                          mock(DistributionLifecycleStateStore.class))
                                                                   .application(APP_ID)
                                                                   .version("1.2.3")
                                                                   .build();
@@ -123,7 +128,8 @@ public class TestAcknowledgingListener {
             AcknowledgingListener listener = AcknowledgingListener.builder()
                                                                   .listener(new Fails())
                                                                   .sink(collector)
-                                                                  .stateStore(mock(DistributionLifecycleStateStore.class))
+                                                                  .stateStore(
+                                                                          mock(DistributionLifecycleStateStore.class))
                                                                   .application(APP_ID)
                                                                   .version("1.2.3")
                                                                   .build();
@@ -146,7 +152,8 @@ public class TestAcknowledgingListener {
             AcknowledgingListener listener = AcknowledgingListener.builder()
                                                                   .listener(new Infinite())
                                                                   .sink(collector)
-                                                                  .stateStore(mock(DistributionLifecycleStateStore.class))
+                                                                  .stateStore(
+                                                                          mock(DistributionLifecycleStateStore.class))
                                                                   .application(APP_ID)
                                                                   .version("1.2.3")
                                                                   .build();
@@ -176,20 +183,20 @@ public class TestAcknowledgingListener {
         File stateFile = Files.createTempFile("state", ".json").toFile();
         stateFile.delete();
         try (DistributionLifecycleStateStore stateStore = AppDistributionLifecycleStoreFile.builder()
-                                                                                      .app(APP_ID)
-                                                                                      .stateFile(stateFile)
-                                                                                      .build()) {
+                                                                                           .app(APP_ID)
+                                                                                           .stateFile(stateFile)
+                                                                                           .build()) {
             stateStore.add(action);
             try (CollectorSink<Event<UUID, LazyEnvelope>> collector = CollectorSink.of()) {
                 AcknowledgingListener listener = AcknowledgingListener.builder()
                                                                       .listener(new TemporarilyFails(1))
                                                                       .sink(e -> {
-                                                            LifecycleAcknowledgement ack =
-                                                                    e.value().getValue().getBodyAs(
-                                                                            LifecycleAcknowledgement.class);
-                                                            stateStore.add(APP_ID, ack);
-                                                            collector.send(e);
-                                                        })
+                                                                          LifecycleAcknowledgement ack =
+                                                                                  e.value().getValue().getBodyAs(
+                                                                                          LifecycleAcknowledgement.class);
+                                                                          stateStore.add(APP_ID, ack);
+                                                                          collector.send(e);
+                                                                      })
                                                                       .stateStore(stateStore)
                                                                       .application(APP_ID)
                                                                       .version("1.2.3")
@@ -202,6 +209,82 @@ public class TestAcknowledgingListener {
                 // Then
                 verifyAcks(collector, ApplicationState.Requested, ApplicationState.InProgress, ApplicationState.Failed,
                            ApplicationState.InProgress, ApplicationState.Completed);
+            }
+        }
+    }
+
+    @Test
+    public void givenAckingListener_whenReceivingActionAlreadyCompleted_thenNoAcksSent_andInnerListenerNeverCalled() throws
+            IOException {
+        // Given
+        LifecycleAction action = action(UUID.randomUUID(), "distro", DistributionLifecycleState.Registered,
+                                        DistributionLifecycleState.Active);
+        File stateFile = Files.createTempFile("state", ".json").toFile();
+        stateFile.delete();
+        try (DistributionLifecycleStateStore stateStore = AppDistributionLifecycleStoreFile.builder()
+                                                                                           .app(APP_ID)
+                                                                                           .stateFile(stateFile)
+                                                                                           .build()) {
+            stateStore.add(action);
+            stateStore.add(APP_ID, ack(action.getEventId(), DISTRO_ID, ApplicationState.Requested));
+            stateStore.add(APP_ID, ack(action.getEventId(), DISTRO_ID, ApplicationState.InProgress));
+            stateStore.add(APP_ID, ack(action.getEventId(), DISTRO_ID, ApplicationState.Completed));
+
+            // When
+            DistributionLifecycleListener listener = mock(DistributionLifecycleListener.class);
+            Sink<Event<UUID, LazyEnvelope>> sink = mock(Sink.class);
+            try (AcknowledgingListener acknowledgingListener = AcknowledgingListener.builder()
+                                                                                    .application(APP_ID)
+                                                                                    .listener(listener)
+                                                                                    .sink(sink)
+                                                                                    .stateStore(stateStore)
+                                                                                    .version("1.2.3")
+                                                                                    .build()) {
+                acknowledgingListener.accept(action);
+
+                // Then
+                verifyNoInteractions(sink);
+
+                // And
+                verifyNoInteractions(listener);
+            }
+        }
+    }
+
+    @Test
+    public void givenAckingListener_whenReceivingActionAlreadyInProgress_thenPartialAcksSent_andInnerListenerCalled() throws
+            IOException {
+        // Given
+        LifecycleAction action = action(UUID.randomUUID(), "distro", DistributionLifecycleState.Registered,
+                                        DistributionLifecycleState.Active);
+        File stateFile = Files.createTempFile("state", ".json").toFile();
+        stateFile.delete();
+        try (DistributionLifecycleStateStore stateStore = AppDistributionLifecycleStoreFile.builder()
+                                                                                           .app(APP_ID)
+                                                                                           .stateFile(stateFile)
+                                                                                           .build()) {
+            stateStore.add(action);
+            stateStore.add(APP_ID, ack(action.getEventId(), DISTRO_ID, ApplicationState.Requested));
+            stateStore.add(APP_ID, ack(action.getEventId(), DISTRO_ID, ApplicationState.InProgress));
+
+            // When
+            DistributionLifecycleListener listener = mock(DistributionLifecycleListener.class);
+            try (CollectorSink<Event<UUID, LazyEnvelope>> sink = CollectorSink.of()) {
+                try (AcknowledgingListener acknowledgingListener = AcknowledgingListener.builder()
+                                                                                        .application(APP_ID)
+                                                                                        .listener(listener)
+                                                                                        .sink(sink)
+                                                                                        .stateStore(stateStore)
+                                                                                        .version("1.2.3")
+                                                                                        .build()) {
+                    acknowledgingListener.accept(action);
+
+                    // Then
+                    verifyAcks(sink, ApplicationState.InProgress, ApplicationState.Completed);
+
+                    // And
+                    verify(listener).accept(action);
+                }
             }
         }
     }